Add captions, force height thumbnail, use avif if availablejo
Return an object to the frontend instead of just list of object ids to
fetch. Set caption if one is set in R2 custom metadata. Otherwise,
generate one from image name (which should be the date).

Also force thumbnails to be 150x199 for consistency and use avif/webp if
available in the browser.